+++ test/dummy/app/controllers/home_controller.rb
@@ -1,4 +1,6 @@
 class HomeController < ApplicationController
+  layout "sidebar"
+
   def index
   end
 end